requirements.txt
python-dotenv
PyYAML
cx-Oracle
SQLAlchemy
playwright
pywin32